+++ packages/terra-disclosure-manager/src/DisclosureManager.jsx
@@ -13,37 +13,16 @@ const availableDisclosureSizes = {
   FULLSCREEN: 'fullscreen',
 };
 
-const availableDisclosureHeights = {
-  240: 240,
-  420: 420,
-  600: 600,
-  690: 690,
-  780: 780,
-  870: 870,
-  960: 960,
-  1140: 1140,
-};
-
-const availableDisclosureWidths = {
-  320: 320,
-  480: 480,
-  640: 640,
-  800: 800,
-  960: 960,
-  1120: 1120,
-  1280: 1280,
-  1440: 1440,
-  1600: 1600,
-  1760: 1760,
-  1920: 1920,
-};
+const arrayReducer = (mappingObject, value) => Object.assign({ [`${value}`]: value }, mappingObject);
+const availableDisclosureHeights = [240, 420, 600, 690, 780, 870, 960, 1140].reduce(arrayReducer, {});
+const availableDisclosureWidths = [320, 480, 640, 800, 960, 1120, 1280, 1440, 1600, 1760, 1920].reduce(arrayReducer, {});
 
 const defaultDimensions = { height: availableDisclosureHeights['690'], width: availableDisclosureWidths['1120'] };
 const defaultSize = availableDisclosureSizes.SMALL;
 
 const isValidDimensions = dimensions => availableDisclosureHeights[dimensions.height] && availableDisclosureWidths[dimensions.width];
 
-const isValidSize = size => availableDisclosureSizes[size.toUpperCase()];
+const isValidSize = size => !!availableDisclosureSizes[size.toUpperCase()];
 
 export { availableDisclosureSizes, availableDisclosureHeights, availableDisclosureWidths };
 
@@ -80,6 +59,7 @@ class DisclosureManager extends React.Component {
     const newState = Object.assign({}, state);
     newState.disclosureComponentKeys = Object.assign([], newState.disclosureComponentKeys);
     newState.disclosureComponentData = Object.assign({}, newState.disclosureComponentData);
+    newState.disclosureComponents = Object.assign([], newState.disclosureComponents);
 
     return newState;
   }
@@ -111,14 +91,45 @@ class DisclosureManager extends React.Component {
     this.onDismissResolvers = {};
 
     this.state = {
+      childComponents: this.renderContentComponents(props.disclosureManager, props.children),
       disclosureIsOpen: false,
-      disclosureIsFocused: true,
+      disclosureIsFocused: false,
       disclosureIsMaximized: false,
       disclosureSize: undefined,
       disclosureDimensions: undefined,
       disclosureComponentKeys: [],
       disclosureComponentData: {},
+      disclosureComponents: [],
     };
+  }
+
+  componentDidUpdate(prevProps) {
+    const { children, disclosureManager } = this.props;
+
+    if (children !== prevProps.children || disclosureManager !== prevProps.disclosureManager) {
+      /**
+       * With the future deprecation of componentWillReceiveProps, there is no existing lifecycle method
+       * that allows for old and new prop comparisons before rendering. componentDidUpdate can be
+       * used to ensure that the state is sycned with the props when necessary; however, if the state does
+       * need to be updated, a second render will need to occur. I do not anticipate this happening often
+       * given how the component is typically consumed.
+       *
+       * Usage of setState in componentDidUpdate is flagged by eslint due to the potential of infinite renders.
+       * However, this implementation of setState should be safe.
+       */
+      // eslint-disable-next-line react/no-did-update-set-state
+      this.setState({
+        childComponents: this.renderContentComponents(disclosureManager, children),
+        disclosureIsOpen: false,
+        disclosureIsFocused: false,
+        disclosureIsMaximized: false,
+        disclosureSize: undefined,
+        disclosureDimensions: undefined,
+        disclosureComponentKeys: [],
+        disclosureComponentData: {},
+        disclosureComponents: [],
+      });
+    }
   }
 
   /**
@@ -150,7 +161,7 @@ class DisclosureManager extends React.Component {
       size = defaultSize;
     }
 
-    this.setState({
+    const newState = {
       disclosureIsOpen: true,
       disclosureIsFocused: true,
       disclosureSize: size,
@@ -164,7 +175,9 @@ class DisclosureManager extends React.Component {
           component: data.content.component,
         },
       },
-    });
+    };
+
+    this.setState(Object.assign(newState, { disclosureComponents: this.renderDisclosureComponents(this.props.disclosureManager, newState) }));
   }
 
   pushDisclosure(data) {
@@ -177,6 +190,7 @@ class DisclosureManager extends React.Component {
       props: data.content.props,
       component: data.content.component,
     };
+    newState.disclosureComponents = newState.disclosureComponents.concat(this.renderDisclosureComponents(this.props.disclosureManager, newState, [data.content.key]));
 
     this.setState(newState);
   }
@@ -185,6 +199,7 @@ class DisclosureManager extends React.Component {
     const newState = DisclosureManager.cloneDisclosureState(this.state);
 
     newState.disclosureComponentData[newState.disclosureComponentKeys.pop()] = undefined;
+    newState.disclosureComponents.pop();
 
     this.setState(newState);
   }
@@ -198,31 +213,36 @@ class DisclosureManager extends React.Component {
       disclosureDimensions: undefined,
       disclosureComponentKeys: [],
       disclosureComponentData: {},
+      disclosureComponents: [],
     });
   }
 
   requestDisclosureFocus() {
-    this.setState({
-      disclosureIsFocused: true,
-    });
+    const newState = DisclosureManager.cloneDisclosureState(this.state);
+    newState.disclosureIsFocused = true;
+
+    this.setState(Object.assign(newState, { disclosureComponents: this.renderDisclosureComponents(this.props.disclosureManager, newState) }));
   }
 
   releaseDisclosureFocus() {
-    this.setState({
-      disclosureIsFocused: false,
-    });
+    const newState = DisclosureManager.cloneDisclosureState(this.state);
+    newState.disclosureIsFocused = false;
+
+    this.setState(Object.assign(newState, { disclosureComponents: this.renderDisclosureComponents(this.props.disclosureManager, newState) }));
   }
 
   maximizeDisclosure() {
-    this.setState({
-      disclosureIsMaximized: true,
-    });
+    const newState = DisclosureManager.cloneDisclosureState(this.state);
+    newState.disclosureIsMaximized = true;
+
+    this.setState(Object.assign(newState, { disclosureComponents: this.renderDisclosureComponents(this.props.disclosureManager, newState) }));
   }
 
   minimizeDisclosure() {
-    this.setState({
-      disclosureIsMaximized: false,
-    });
+    const newState = DisclosureManager.cloneDisclosureState(this.state);
+    newState.disclosureIsMaximized = false;
+
+    this.setState(Object.assign(newState, { disclosureComponents: this.renderDisclosureComponents(this.props.disclosureManager, newState) }));
   }
 
   /**
@@ -301,9 +321,14 @@ class DisclosureManager extends React.Component {
     };
   }
 
-  renderContentComponents() {
-    const { children, disclosureManager } = this.props;
-
+  /**
+   * Generates an Array of component instances based on the given children. These components can be stored in state to ensure subsequent
+   * renders are more efficient and do not unnecessarily generate new DisclosureManagerDelegate instances.
+   * @param {DisclosureManagerDelegate} disclosureManager A DisclosureManagerDelegate instance that will be used as a fallback in cases where unsupported
+   *                                                      disclosure types are requested.
+   * @param {Children} children The children prop value that will be wrapped in the generated provider.
+   */
+  renderContentComponents(disclosureManager, children) {
     const delegate = {};
 
     /**
@@ -343,16 +368,25 @@ class DisclosureManager extends React.Component {
     );
   }
 
-  renderDisclosureComponents() {
-    const { disclosureManager } = this.props;
+  /**
+   * Generates an Array of component instances based on the given disclosure state. These components can be stored in state to ensure subsequent
+   * renders are more efficient and do not unnecessarily generate new DisclosureManagerDelegate instances.
+   * @param {DisclosureManagerDelegate} disclosureManager A DisclosureManagerDelegate instance that will be used as a fallback in cases where unsupported
+   *                                                      disclosure types are requested.
+   * @param {Object} disclosureState An Object representing the state of the disclosure manager.
+   * @param {Array} componentKeysOverride An Array of component keys representing the components that are to be generated. If not provided, all component keys
+   *                                      found in the disclosureState will be used for generation.
+   */
+  renderDisclosureComponents(disclosureManager, disclosureState, componentKeysOverride) {
     const {
       disclosureComponentKeys, disclosureComponentData, disclosureIsMaximized, disclosureIsFocused, disclosureSize,
-    } = this.state;
+    } = disclosureState;
 
-    return disclosureComponentKeys.map((componentKey, index) => {
+    return (componentKeysOverride || disclosureComponentKeys).map((componentKey) => {
       const componentData = disclosureComponentData[componentKey];
       const isFullscreen = disclosureSize === availableDisclosureSizes.FULLSCREEN;
       const popContent = this.generatePopFunction(componentData.key);
+      const componentIndex = disclosureComponentKeys.indexOf(componentKey);
 
       const delegate = {};
 
@@ -380,7 +414,7 @@ class DisclosureManager extends React.Component {
        * Allows a component to remove itself from the disclosure stack. If the component is the only element in the disclosure stack,
        * the disclosure is closed.
        */
-      delegate.dismiss = index > 0 ? popContent : this.safelyCloseDisclosure;
+      delegate.dismiss = componentIndex > 0 ? popContent : this.safelyCloseDisclosure;
 
       /**
        * Allows a component to close the entire disclosure stack.
@@ -391,7 +425,7 @@ class DisclosureManager extends React.Component {
        * Allows a component to remove itself from the disclosure stack. Functionally similar to `dismiss`, however `onBack` is
        * only provided to components in the stack that have a previous sibling.
        */
-      delegate.goBack = index > 0 ? popContent : undefined;
+      delegate.goBack = componentIndex > 0 ? popContent : undefined;
 
       /**
        * Allows a component to request focus from the disclosure in the event that the disclosure mechanism in use utilizes a focus trap.
@@ -429,11 +463,9 @@ class DisclosureManager extends React.Component {
       };
 
       return (
-        <React.Fragment key={componentData.key}>
-          <DisclosureManagerContext.Provider value={DisclosureManagerDelegate.create(delegate)}>
-            {componentData.component}
-          </DisclosureManagerContext.Provider>
-        </React.Fragment>
+        <DisclosureManagerContext.Provider value={DisclosureManagerDelegate.create(delegate)} key={componentData.key}>
+          {componentData.component}
+        </DisclosureManagerContext.Provider>
       );
     });
   }
@@ -441,12 +473,14 @@ class DisclosureManager extends React.Component {
   render() {
     const { render } = this.props;
     const {
+      childComponents,
       disclosureIsOpen,
       disclosureIsFocused,
       disclosureIsMaximized,
       disclosureSize,
       disclosureDimensions,
       disclosureComponentKeys,
+      disclosureComponents,
     } = this.state;
 
     if (!render) {
@@ -457,7 +491,7 @@ class DisclosureManager extends React.Component {
       dismissPresentedComponent: (disclosureComponentKeys.length > 1) ? this.generatePopFunction(disclosureComponentKeys[disclosureComponentKeys.length - 1]) : this.safelyCloseDisclosure,
       closeDisclosure: this.safelyCloseDisclosure,
       children: {
-        components: this.renderContentComponents(),
+        components: childComponents,
       },
       disclosure: {
         isOpen: disclosureIsOpen,
@@ -465,7 +499,7 @@ class DisclosureManager extends React.Component {
         isMaximized: disclosureIsMaximized,
         size: disclosureSize,
         dimensions: disclosureDimensions,
-        components: this.renderDisclosureComponents(),
+        components: disclosureComponents,
       },
     });
   }